+++ src/core/explicit-dynamics.cpp
@@ -1,9 +1,8 @@
-#include "aligator/core/explicit-dynamics.hxx"
+#include "aligator/core/explicit-dynamics.hpp"
 
 namespace aligator {
 
 template struct ExplicitDynamicsModelTpl<context::Scalar>;
-
 template struct ExplicitDynamicsDataTpl<context::Scalar>;
 
 } // namespace aligator